Move resp components into their own folder
